% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_data}
\alias{fake_data}
\title{Fake data}
\format{
A data frame with 28 variables and 3000 random obs.
}
\usage{
fake_data
}
\description{
A simulated fake data set created by Xin Xia.
}
\keyword{datasets}
