% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_snack_df}
\alias{fake_snack_df}
\title{Fake SNAC-K data}
\format{
A data frame with 15 variables and 3365 random obs.
}
\usage{
fake_snack_df
}
\description{
A randomized SNAC-K data set, with metadata preserved.
}
\details{
Random values are generated by \code{rnorm()} and \code{sample()}.
}
\keyword{datasets}
