% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code_2023.R
\name{get_code_2023}
\alias{get_code_2023}
\title{Get R code for workshop in year 2023}
\usage{
get_code_2023(session = NULL, open = interactive(), ...)
}
\arguments{
\item{session}{Either 'loop' or 'tidyverse'.}

\item{open}{should the file be opened after being created}

\item{...}{arguments to be passed to \link[usethis]{use_template}}
}
\description{
This function creates a new folder "R_example_code" at root of project and create an R script file.

Author: Xin Xia (loop) and Bolin Wu (tidyverse)
}
\examples{
\dontrun{
workshopr::get_code_2023(session = "loop")
workshopr::get_code_2023(session = "tidyverse")
}

}
