% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rmd_2023.R
\name{get_rmd_2023}
\alias{get_rmd_2023}
\title{get_rmd_2023}
\usage{
get_rmd_2023(
  name = NULL,
  subDir = "rmd",
  output_file = "html",
  open = interactive(),
  ...
)
}
\arguments{
\item{name}{Name of rmd file. Please do not use '/' in the the file name (no need to add suffix ".Rmd").}

\item{subDir}{Name of the folder where rmd file will be created. Remember to connect sub-dir names with '/'. By default it is 'rmd'.}

\item{output_file}{"word" or "html" or "pdf", by default it is "html".}

\item{open}{Should the file be opened after being created.}

\item{...}{Arguments to be passed to \link[usethis]{use_template}}
}
\description{
Use the preset rmd template. By default, this function creates a new folder "rmd" at root of project and create an rmd file with nice-looking template there.
}
\examples{
\dontrun{
workshopr::get_rmd_2023(subDir = "rmd", name = "pretty_template", output_file = "word")
}

}
