% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solution_2023.R
\name{get_solution_2023}
\alias{get_solution_2023}
\title{get_solution_2023}
\usage{
get_solution_2023(name = NULL, subDir = "rmd", open = interactive(), ...)
}
\arguments{
\item{name}{Name of rmd file. Please do not use '/' in the the file name (no need to add suffix ".Rmd").}

\item{subDir}{Name of the folder where rmd file will be created. Remember to connect sub-dir names with '/'. By default it is 'rmd'.}

\item{open}{Should the file be opened after being created.}

\item{...}{Arguments to be passed to \link[usethis]{use_template}}
}
\description{
Get the solution of final practice.

#' Author: Xin Xia
}
\examples{
\dontrun{
workshopr::get_solution_2023(name = "practice_solution")
}


}
