% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_session.R
\name{loop_session_2023}
\alias{loop_session_2023}
\title{Loop session code}
\usage{
loop_session_2023(open = interactive(), ...)
}
\arguments{
\item{open}{should the file be opened after being created}

\item{...}{arguments to be passed to \link[usethis]{use_template}}
}
\description{
Load the R code for loop session at R workshop. Author: Xin Xia.
}
\examples{
\dontrun{
workshopr::loop_session_2023()
}

}
