% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paquid_cog}
\alias{paquid_cog}
\title{paquid: cognitive functions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2250 rows and 8 columns.
}
\source{
\url{https://www.rdocumentation.org/packages/lcmm/versions/1.8.1.1/topics/paquid}
}
\usage{
paquid_cog
}
\description{
Can use this data to check the cognitive funciton in paquid dataset.
}
\examples{
  # load data
  paquid_cog
}
\keyword{datasets}
