% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paquid_cov}
\alias{paquid_cov}
\title{paquid: physical covariance}
\format{
Data frame with columns
}
\source{
\url{https://www.rdocumentation.org/packages/lcmm/versions/1.8.1.1/topics/paquid}
}
\usage{
paquid_cov
}
\description{
Can use this data to check the covariance of physical funciton decrease in paquid funciton.
}
\examples{
  # load data
  paquid_cov


}
\keyword{datasets}
