% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paquid_cov}
\alias{paquid_cov}
\title{paquid: factors of cognition decrease}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2250 rows and 7 columns.
}
\source{
\url{https://www.rdocumentation.org/packages/lcmm/versions/1.8.1.1/topics/paquid}
}
\usage{
paquid_cov
}
\description{
Factors of cognition decrease in paquid dataset.
}
\examples{
\dontrun{
  # load data
  workshopr::paquid_cov
}

}
\keyword{datasets}
